fix: using AtomicUsize instead of AtomicU64

+++ src/kanal_tests.rs
@@ -6,7 +6,7 @@ const MESSAGES: usize = 1024;
 const THREADS: usize = 8;
 
 use std::sync::{
-    atomic::{AtomicU64, Ordering},
+    atomic::{AtomicUsize, Ordering},
     Arc,
 };
 
@@ -28,7 +28,7 @@ struct PaddedReprC {
 struct DropTester {
     i: usize,
     dropped: bool,
-    counter: Arc<AtomicU64>,
+    counter: Arc<AtomicUsize>,
 }
 
 impl Drop for DropTester {
@@ -46,7 +46,7 @@ impl Drop for DropTester {
 
 impl DropTester {
     #[allow(dead_code)]
-    fn new(counter: Arc<AtomicU64>, i: usize) -> Self {
+    fn new(counter: Arc<AtomicUsize>, i: usize) -> Self {
         if i == 0 {
             panic!("don't initialize DropTester with 0");
         }
@@ -62,7 +62,7 @@ impl DropTester {
 mod tests {
     use crate::kanal_tests::*;
     use crate::{bounded, unbounded, ReceiveError, Receiver, SendError, Sender};
-    use std::sync::atomic::{AtomicU64, Ordering};
+    use std::sync::atomic::{AtomicUsize, Ordering};
     use std::sync::Arc;
     use std::time::Duration;
 
@@ -312,44 +312,44 @@ mod tests {
     // Channel drop tests
     #[test]
     fn drop_test() {
-        let counter = Arc::new(AtomicU64::new(0));
+        let counter = Arc::new(AtomicUsize::new(0));
         mpmc_dyn!(DropTester::new(counter.clone(), 10), Some(1));
-        assert_eq!(counter.load(Ordering::SeqCst), MESSAGES as u64);
+        assert_eq!(counter.load(Ordering::SeqCst), MESSAGES as usize);
     }
 
     #[test]
     fn drop_test_in_queue() {
-        let counter = Arc::new(AtomicU64::new(0));
+        let counter = Arc::new(AtomicUsize::new(0));
         let (s, r) = new(Some(10));
         for _ in 0..10 {
             s.send(DropTester::new(counter.clone(), 1234)).unwrap();
         }
         r.close();
-        assert_eq!(counter.load(Ordering::SeqCst), 10_u64);
+        assert_eq!(counter.load(Ordering::SeqCst), 10_usize);
     }
 
     #[test]
     fn drop_test_send_to_closed() {
-        let counter = Arc::new(AtomicU64::new(0));
+        let counter = Arc::new(AtomicUsize::new(0));
         let (s, r) = new(Some(10));
         r.close();
         for _ in 0..10 {
             // will fail
             let _ = s.send(DropTester::new(counter.clone(), 1234));
         }
-        assert_eq!(counter.load(Ordering::SeqCst), 10_u64);
+        assert_eq!(counter.load(Ordering::SeqCst), 10_usize);
     }
 
     #[test]
     fn drop_test_send_to_half_closed() {
-        let counter = Arc::new(AtomicU64::new(0));
+        let counter = Arc::new(AtomicUsize::new(0));
         let (s, r) = new(Some(10));
         drop(r);
         for _ in 0..10 {
             // will fail
             let _ = s.send(DropTester::new(counter.clone(), 1234));
         }
-        assert_eq!(counter.load(Ordering::SeqCst), 10_u64);
+        assert_eq!(counter.load(Ordering::SeqCst), 10_usize);
     }
 
     #[test]
@@ -357,7 +357,7 @@ mod tests {
         let (s, r) = new(Some(0));
 
         crossbeam::scope(|scope| {
-            let counter = Arc::new(AtomicU64::new(0));
+            let counter = Arc::new(AtomicUsize::new(0));
             let mut list = Vec::new();
             for _ in 0..10 {
                 let counter = counter.clone();
@@ -372,7 +372,7 @@ mod tests {
             for t in list {
                 t.join().unwrap();
             }
-            assert_eq!(counter.load(Ordering::SeqCst), 10_u64);
+            assert_eq!(counter.load(Ordering::SeqCst), 10_usize);
         })
         .unwrap();
     }
@@ -443,7 +443,7 @@ mod async_tests {
         bounded_async, unbounded_async, AsyncReceiver, AsyncSender, ReceiveError, SendError,
     };
 
-    use std::sync::atomic::{AtomicU64, Ordering};
+    use std::sync::atomic::{AtomicUsize, Ordering};
     use std::sync::Arc;
     use std::time::Duration;
 
@@ -642,16 +642,16 @@ mod async_tests {
 
     #[tokio::test]
     async fn async_drop_test() {
-        let counter = Arc::new(AtomicU64::new(0));
+        let counter = Arc::new(AtomicUsize::new(0));
         async_mpmc_dyn!(let counter=counter.clone(),DropTester::new(counter.clone(), 10), Some(1));
-        assert_eq!(counter.load(Ordering::SeqCst), MESSAGES as u64);
+        assert_eq!(counter.load(Ordering::SeqCst), MESSAGES as usize);
     }
 
     #[tokio::test]
     async fn async_drop_test_in_signal() {
         let (s, r) = new_async(Some(10));
 
-        let counter = Arc::new(AtomicU64::new(0));
+        let counter = Arc::new(AtomicUsize::new(0));
         let mut list = Vec::new();
         for _ in 0..10 {
             let counter = counter.clone();
@@ -665,7 +665,7 @@ mod async_tests {
         for c in list {
             c.await.unwrap();
         }
-        assert_eq!(counter.load(Ordering::SeqCst), 10_u64);
+        assert_eq!(counter.load(Ordering::SeqCst), 10_usize);
     }
 
     // Channel logic tests
@@ -741,7 +741,7 @@ mod async_tests {
     #[tokio::test]
     async fn async_send_abort_test() {
         let (s, r) = new_async::<DropTester>(Some(0));
-        let counter = Arc::new(AtomicU64::new(0));
+        let counter = Arc::new(AtomicUsize::new(0));
         let mut list = Vec::new();
         for _ in 0..10 {
             let s = s.clone();
@@ -758,7 +758,7 @@ mod async_tests {
             c.abort();
         }
         tokio::time::sleep(Duration::from_millis(500)).await;
-        assert_eq!(counter.load(Ordering::SeqCst), 10_u64);
+        assert_eq!(counter.load(Ordering::SeqCst), 10_usize);
         r.close();
     }
 
@@ -766,7 +766,7 @@ mod async_tests {
     async fn async_drop_test_in_queue() {
         let (s, r) = new_async(Some(10));
 
-        let counter = Arc::new(AtomicU64::new(0));
+        let counter = Arc::new(AtomicUsize::new(0));
         let mut list = Vec::new();
         for _ in 0..10 {
             let counter = counter.clone();
@@ -780,44 +780,44 @@ mod async_tests {
             c.await.unwrap();
         }
         r.close();
-        assert_eq!(counter.load(Ordering::SeqCst), 10_u64);
+        assert_eq!(counter.load(Ordering::SeqCst), 10_usize);
     }
 
     #[tokio::test]
     async fn async_drop_test_in_unused_signal() {
         let (s, r) = new_async(Some(10));
 
-        let counter = Arc::new(AtomicU64::new(0));
+        let counter = Arc::new(AtomicUsize::new(0));
         for _ in 0..10 {
             let counter = counter.clone();
             let _ = s.send(DropTester::new(counter, 1234));
         }
         r.close();
-        assert_eq!(counter.load(Ordering::SeqCst), 10_u64);
+        assert_eq!(counter.load(Ordering::SeqCst), 10_usize);
     }
 
     #[tokio::test]
     async fn async_drop_test_send_to_closed() {
         let (s, r) = new_async(Some(10));
         r.close();
-        let counter = Arc::new(AtomicU64::new(0));
+        let counter = Arc::new(AtomicUsize::new(0));
         for _ in 0..10 {
             let counter = counter.clone();
             let _ = s.send(DropTester::new(counter, 1234)).await;
         }
-        assert_eq!(counter.load(Ordering::SeqCst), 10_u64);
+        assert_eq!(counter.load(Ordering::SeqCst), 10_usize);
     }
 
     #[tokio::test]
     async fn async_drop_test_send_to_half_closed() {
         let (s, r) = new_async(Some(10));
         drop(r);
-        let counter = Arc::new(AtomicU64::new(0));
+        let counter = Arc::new(AtomicUsize::new(0));
         for _ in 0..10 {
             let counter = counter.clone();
             let _ = s.send(DropTester::new(counter, 1234)).await;
         }
-        assert_eq!(counter.load(Ordering::SeqCst), 10_u64);
+        assert_eq!(counter.load(Ordering::SeqCst), 10_usize);
     }
 
     #[tokio::test]
